Remove #pragma pack() calls for MSC.
It looks like it was added for the bit-packed structures, but not moved when
srtp_priv.h was split out.  Correct packing of bit-packed structures is now
resolved differently for Microsoft C.

With the previous code, it could result in the structures defined in
crypto_kernel.h and its includes being compiled with inconsistent packing,
depending on the order in which header files are included in C files.